feat(build): Use `swiftc` instead of Xcode for building
Summary:
Xcode sucks.  Especially for a project like this with literally two source
files and no dependencies.

I recommended this tool to a friend who struggled installing this program for
over ten minutes and it's my fault.

This commit removes the Xcode dependency and replaces it with a simple `swift`
command.  This should make it easier to install and use.

APODesktop can still be built with xcode using `make xcode`

Test Plan:

```sh
make
```