fix(windows): don't crash when ran from Task Scheduler
Summary:
Task Scheduler runs programs from C:\System32.  APODesktop saved images in the
current directory which would cause it to crash when ran from Task Scheduler
because System32 is not writable.

This commit teaches APODesktop to write the files to a temporary directory that it owns.

Test plan:
1. Make a Task Scheduler thingy for APODesktop
2. wait for it to run

before this commit:
crashed

this commit:
no crash