feat(windows): Don't open cmd when running
Summary:
Previously, when Task Scheduler would run this program it would open a cmd
window.  This made it annoying because even though it'd only show up for a
moment it would steal focus.

This commit changes APODesktop from a Console app to a Windows app.

Test Plan:
1. Set APODesktop's schedule to be in 1 minute from now
2. wait for it to run

Previous commit: it will open a cmd window.
This commit: nothing is noticeable from the user's perspective.


Thanks to @strager for the suggestion!